Use rbenv instead of chruby; fix PATH
* Remove sourced chruby scripts
* Update `PATH` for use with rbenv

Note: The additions to `PATH` here are also found in [`.zshenv`][zshenv]
in [thoughtbot/dotfiles], which is sourced before the `.zshrc.local` in
this repository. However, OS X's `path_helper` in `/etc/zshrc`, which is
sourced *after* `.zshenv`, messes with `PATH`. Specifically, the OS X
`path_helper` [prepends] the contents of `/etc/paths` (and others) to
anything that's already in `PATH`. In order for rbenv to work properly,
we have to re-prepend its shims to `PATH` after the `path_helper` runs,
so we do so here in `.zshrc.local`, despite the duplication.

[zshenv]: https://github.com/thoughtbot/dotfiles/blob/master/zshenv
[thoughtbot/dotfiles]: https://github.com/thoughtbot/dotfiles
[prepends]: http://stackoverflow.com/a/14101578