[vim] Don't change directory when linting python
ALE now runs `flake8` and `pylint` in the directory containing the file
being linted. This breaks the ability to set a custom linter config file
which is relative to the buffer directory.

The added config values in this commit tells ALE not to change to the
file's directory before running these linters.