[git] Add `git fixup` command
When working on a feature branch, I sometimes find that I want to amend
a commit which is not `HEAD`. This requires committing a `fixup!` commit
with the changes and performing an interactive rebase.

This script adds a `git fixup` command which creates the `!fixup` commit
and immediately squashes it.

The script ensures you do not attempt to perform a rebase of a commit on
the master branch.

Based on: https://blog.filippo.io/git-fixup-amending-an-older-commit/